+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/i18n/DefaultExternalTranslationProvider.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.common.service.impl;
+package org.opentestsystem.rdw.reporting.i18n;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
@@ -19,7 +19,7 @@ import java.util.Locale;
  * This class provides translations from the configured {@link TenantProperties#getTranslationLocation()}.
  */
 @Component
-public class DefaultExternalTranslationProvider implements TranslationProvider {
+class DefaultExternalTranslationProvider implements TranslationProvider {
     private final static Logger logger = LoggerFactory.getLogger(DefaultExternalTranslationProvider.class);
 
     private final TenantProperties tenantProperties;
@@ -27,7 +27,7 @@ public class DefaultExternalTranslationProvider implements TranslationProvider {
     private final ObjectMapper objectMapper;
 
     @Autowired
-    public DefaultExternalTranslationProvider(final TenantProperties tenantProperties,
+    DefaultExternalTranslationProvider(final TenantProperties tenantProperties,
                                               final ResourceLoader resourceLoader,
                                               final ObjectMapper objectMapper) {
         this.tenantProperties = tenantProperties;
