+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/i18n/DefaultFrontendRepositoryTranslationProvider.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.common.service.impl;
+package org.opentestsystem.rdw.reporting.i18n;
 
 import org.opentestsystem.rdw.reporting.common.i18n.AbstractRepositoryTranslationProvider;
 import org.opentestsystem.rdw.reporting.common.i18n.TranslationRepository;
@@ -11,10 +11,10 @@ import static org.opentestsystem.rdw.reporting.common.i18n.TranslationNamespace.
  * This translation provider provides repository messages for the FRONTEND namespace.
  */
 @Component
-public class DefaultFrontendRepositoryTranslationProvider extends AbstractRepositoryTranslationProvider {
+class DefaultFrontendRepositoryTranslationProvider extends AbstractRepositoryTranslationProvider {
 
     @Autowired
-    public DefaultFrontendRepositoryTranslationProvider(final TranslationRepository translationRepository) {
+    DefaultFrontendRepositoryTranslationProvider(final TranslationRepository translationRepository) {
         super(FRONTEND, translationRepository);
     }
