+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/i18n/DefaultTranslationMergeService.java
@@ -1,11 +1,10 @@
-package org.opentestsystem.rdw.reporting.common.service.impl;
+package org.opentestsystem.rdw.reporting.i18n;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.collect.Ordering;
 import org.opentestsystem.rdw.reporting.common.i18n.TranslationProvider;
-import org.opentestsystem.rdw.reporting.common.service.TranslationMergeService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,7 +20,7 @@ import java.util.Set;
  * This will read messages from registered {@link TranslationProvider} instances.
  */
 @Service
-public class DefaultTranslationMergeService implements TranslationMergeService {
+class DefaultTranslationMergeService implements TranslationMergeService {
     private final static Logger logger = LoggerFactory.getLogger(DefaultTranslationMergeService.class);
     private final static Ordering<TranslationProvider> ProviderOrdering = Ordering
             .natural()
@@ -31,7 +30,7 @@ public class DefaultTranslationMergeService implements TranslationMergeService {
     private final ObjectMapper objectMapper;
 
     @Autowired
-    public DefaultTranslationMergeService(final Set<TranslationProvider> translationProviders,
+    DefaultTranslationMergeService(final Set<TranslationProvider> translationProviders,
                                           final ObjectMapper objectMapper) {
         this.translationProviders = ProviderOrdering.sortedCopy(translationProviders);
         this.objectMapper = objectMapper;
