+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/i18n/TranslationController.java
@@ -1,7 +1,6 @@
-package org.opentestsystem.rdw.reporting.common.web;
+package org.opentestsystem.rdw.reporting.i18n;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import org.opentestsystem.rdw.reporting.common.service.TranslationMergeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,17 +13,19 @@ import java.util.Locale;
  * from both the database and external xx.json resources.
  */
 @RestController
-public class TranslationController {
+@RequestMapping("/api/translations")
+class TranslationController {
 
     private final TranslationMergeService service;
 
     @Autowired
-    public TranslationController(final TranslationMergeService service) {
+    TranslationController(final TranslationMergeService service) {
         this.service = service;
     }
 
-    @RequestMapping("/api/translations/{locale}")
+    @RequestMapping("/{locale}")
     public ObjectNode getTranslations(@PathVariable final Locale locale) {
         return service.getLocaleMessages(locale);
     }
+
 }
