+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/i18n/TranslationMergeService.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.common.service;
+package org.opentestsystem.rdw.reporting.i18n;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
