--- webapp/src/main/java/org/opentestsystem/rdw/reporting/i18n/TranslationWebConfiguration.java
@@ -1,19 +1,0 @@
-package org.opentestsystem.rdw.reporting.i18n;
-
-import org.opentestsystem.rdw.reporting.common.service.impl.DefaultExternalTranslationProvider;
-import org.opentestsystem.rdw.reporting.common.service.impl.DefaultFrontendRepositoryTranslationProvider;
-import org.opentestsystem.rdw.reporting.common.service.impl.DefaultTranslationMergeService;
-import org.opentestsystem.rdw.reporting.common.web.TranslationController;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-// TODO move to reporting-service when admin-webapp no longer depends on /api/translations endpoint
-@Configuration
-@Import({
-        DefaultExternalTranslationProvider.class,
-        DefaultFrontendRepositoryTranslationProvider.class,
-        DefaultTranslationMergeService.class,
-        TranslationController.class
-})
-public class TranslationWebConfiguration {
-}
