+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/i18n/DefaultExternalTranslationProviderTest.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.common.service.impl;
+package org.opentestsystem.rdw.reporting.i18n;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
