+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/i18n/TranslationControllerIT.java
@@ -1,11 +1,10 @@
-package org.opentestsystem.rdw.reporting.common.web;
+package org.opentestsystem.rdw.reporting.i18n;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.opentestsystem.rdw.reporting.common.service.TranslationMergeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
